general housekeeping and cleanup
